Create SocialLink component and refactor folder structure in Footer component (4)
