+++ src/components/Footer/index.js
@@ -1,3 +1,5 @@
-export { default as Break } from './Break';
+export { default as Break } from './break/Break';
+export { default as FooterSection } from './footerSection/FooterSection';
+export { default as SocialLink } from './socialLink/SocialLink';
+
 export { default as Footer } from './Footer';
-export { default as FooterSection } from './FooterSection';
