+++ src/components/Footer/socialLink/SocialLink.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function SocialLink({to, Icon}) {
+    return (
+       <a href={to} target="_blank" rel="noopener noreferrer">
+           <Icon className="icon" />
+       </a>
+    );
+}
+
+export default SocialLink;
